ml files/requirements.txt
joblib
pandas
scikit-learn
pyserial
SpeechRecognition
pyaudio
vosk
rich
keyboard
matplotlib
seaborn
numpy
pynput
